Added a configuration option to specify the register.
I thought that it is possible a user may want to specify the register
that the information gets saved in instead of always using `"`.

Now using `g:YankMatches#ClipboardRegister` lets you do this.

Also, now instead of using `"` by default, the script looks at your
`clipboard` setting for the default register to use.